server/requirements.txt
django==4.0.4
djangorestframework==3.13.1
django-cors-headers==3.10.1
psycopg2-binary==2.8.4
